slack_bot/requirements.txt
# TrIAge Slack Bot Dependencies
# Core dependencies
pydantic==2.9.0
httpx==0.27.0
cryptography==42.0.0
slack-sdk==3.27.0
slack-bolt==1.18.0
python-dotenv==1.2.1
asyncpg==0.29.0
